requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
anthropic
httpx
redis
google-api-python-client
google-auth
